backend/requirements.txt
Django==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
